Revert "ChessMove refactoring: inserted bits to save check, checkmate and stalemate information + tests" This reverts commit 0a97db0f6cbc9b1f35de0a57230dd3443befb761.
The last REVERTED commits was done to try to insert "annotation" information
in the ChessMove (check, checkmate, stalemate). The try is aborted because
the insertion of these information is quite complex and made the move search
algorithm too much slower.
Additionally a lot of tests was broken, and try to fix them cause the
code to be messed and still more slower.
The first reason to insert these bits was to try to manage wrong
annotations in PGN moves, frankly speaking a naive task in this
phase of development.